Check module enabled for states migration

+++ src/Database/Migrations/2021_10_19_071730_create_states_table.php
@@ -13,17 +13,19 @@ class CreateStatesTable extends Migration
 	 */
 	public function up()
 	{
-		Schema::create(config('world.migrations.states.table_name'), function (Blueprint $table) {
-			$table->id();
-			$table->foreignId('country_id');
-			$table->string('name');
+		if(config('world.modules.states')){
+			Schema::create(config('world.migrations.states.table_name'), function (Blueprint $table) {
+				$table->id();
+				$table->foreignId('country_id');
+				$table->string('name');
 
-			foreach (config('world.migrations.states.optional_fields') as $field => $value) {
-				if ($value['required']) {
-					$table->string($field, $value['length'] ?? null);
+				foreach (config('world.migrations.states.optional_fields') as $field => $value) {
+					if ($value['required']) {
+						$table->string($field, $value['length'] ?? null);
+					}
 				}
-			}
-		});
+			});
+		}
 	}
 
 	/**
@@ -33,6 +35,8 @@ class CreateStatesTable extends Migration
 	 */
 	public function down()
 	{
-		Schema::dropIfExists(config('world.migrations.states.table_name'));
+		if(config('world.modules.states')){
+			Schema::dropIfExists(config('world.migrations.states.table_name'));
+		}
 	}
 }
